Rename compatibility headers to have cxx14 suffix.
Modify headers and duplicate code (yuk) so that compatibility headers work as
single-file includes in http://godbolt.org
