--- compatibility/in_place_type.h
@@ -1,42 +1,0 @@
-/* Copyright (c) 2016 The Value Types Authors. All Rights Reserved.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-the Software, and to permit persons to whom the Software is furnished to do so,
-subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-==============================================================================*/
-
-#ifndef XYZ_IN_PLACE_TYPE_H
-#define XYZ_IN_PLACE_TYPE_H
-
-#if (__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
-
-#include <utility>
-
-namespace xyz {
-using std::in_place_type_t;
-}  // namespace xyz
-
-#else
-
-namespace xyz {
-template <class T>
-struct in_place_type_t {};
-}  // namespace xyz
-
-#endif  // (__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >=
-        // 201703L)
-
-#endif  // XYZ_IN_PLACE_TYPE_H
