+++ compatibility/polymorphic_cxx14.h
@@ -26,10 +26,15 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include <type_traits>
 #include <utility>
 
-#include "compatibility/empty_base_optimization.h"
-#include "compatibility/in_place_type.h"
-
+#ifndef XYZ_IN_PLACE_TYPE_DEFINED
+#define XYZ_IN_PLACE_TYPE_DEFINED
 namespace xyz {
+
+struct NoPolymorphicSBO {};
+
+template <class T>
+struct in_place_type_t {};
+#endif  // XYZ_IN_PLACE_TYPE_DEFINED
 
 #ifndef XYZ_UNREACHABLE_DEFINED
 #define XYZ_UNREACHABLE_DEFINED
@@ -44,7 +49,38 @@ namespace xyz {
 }
 #endif  // XYZ_UNREACHABLE_DEFINED
 
-struct NoPolymorphicSBO {};
+#ifndef XYZ_EMPTY_BASE_DEFINED
+#define XYZ_EMPTY_BASE_DEFINED
+// This is a helper class to allow empty base class optimization.
+// This implementation is duplicated in compatibility/in_place_type_cxx14.h.
+// These implementations must be kept in sync.
+// We duplicate implementations to allow this header to work as a single
+// include. https://godbolt.org needs single-file includes.
+// TODO: Add tests to keep implementations in sync.
+namespace detail {
+template <class T, bool CanBeEmptyBaseClass =
+                       std::is_empty<T>::value && !std::is_final<T>::value>
+class empty_base_optimization {
+ protected:
+  empty_base_optimization() = default;
+  empty_base_optimization(const T& t) : t_(t) {}
+  empty_base_optimization(T&& t) : t_(std::move(t)) {}
+  T& get() noexcept { return t_; }
+  const T& get() const noexcept { return t_; }
+  T t_;
+};
+
+template <class T>
+class empty_base_optimization<T, true> : private T {
+ protected:
+  empty_base_optimization() = default;
+  empty_base_optimization(const T& t) : T(t) {}
+  empty_base_optimization(T&& t) : T(std::move(t)) {}
+  T& get() noexcept { return *this; }
+  const T& get() const noexcept { return *this; }
+};
+}  // namespace detail
+#endif  // XYZ_EMPTY_BASE_DEFINED
 
 namespace detail {
 template <class T, class A>
